# docs/swagger.yaml
basePath: /api/v1
definitions:
  internal_delivery_http_handler.CreateProductRequest:
    properties:
      description:
        type: string
      name:
        maxLength: 255
        minLength: 1
        type: string
      price:
        minimum: 0
        type: number
    required:
    - name
    - price
    type: object
  internal_delivery_http_handler.CreateReviewRequest:
    properties:
      first_name:
        maxLength: 100
        minLength: 1
        type: string
      last_name:
        maxLength: 100
        minLength: 1
        type: string
      product_id:
        type: string
      rating:
        maximum: 5
        minimum: 1
        type: integer
      review_text:
        minLength: 1
        type: string
    required:
    - first_name
    - last_name
    - product_id
    - rating
    - review_text
    type: object
  internal_delivery_http_handler.UpdateProductRequest:
    properties:
      description:
        type: string
      name:
        maxLength: 255
        minLength: 1
        type: string
      price:
        minimum: 0
        type: number
    required:
    - name
    - price
    type: object
  internal_delivery_http_handler.UpdateReviewRequest:
    properties:
      first_name:
        maxLength: 100
        minLength: 1
        type: string
      last_name:
        maxLength: 100
        minLength: 1
        type: string
      rating:
        maximum: 5
        minimum: 1
        type: integer
      review_text:
        minLength: 1
        type: string
    required:
    - first_name
    - last_name
    - rating
    - review_text
    type: object
host: localhost:8080
info:
  contact:
    email: support@example.com
    name: API Support
    url: http://github.com/Pesokrava/product_reviewer
  description: A production-ready product reviews system with RESTful APIs, caching,
    and event notifications.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: Product Reviews API
  version: "1.0"
paths:
  /products:
    get:
      consumes:
      - application/json
      description: Get a paginated list of products
      parameters:
      - default: 20
        description: Number of items per page (max 100)
        in: query
        name: limit
        type: integer
      - default: 0
        description: Number of items to skip
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Paginated list of products
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: List all products
      tags:
      - Products
    post:
      consumes:
      - application/json
      description: Create a new product with name, description, and price
      parameters:
      - description: Product details
        in: body
        name: product
        required: true
        schema:
          $ref: '#/definitions/internal_delivery_http_handler.CreateProductRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Product created successfully
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid request body
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Create a new product
      tags:
      - Products
  /products/{id}:
    delete:
      consumes:
      - application/json
      description: Soft delete a product and all its reviews
      parameters:
      - description: Product ID (UUID)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: Product deleted successfully
        "400":
          description: Invalid product ID
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Product not found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Delete a product
      tags:
      - Products
    get:
      consumes:
      - application/json
      description: Get detailed information about a product including average rating
      parameters:
      - description: Product ID (UUID)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Product details
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid product ID
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Product not found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get a product by ID
      tags:
      - Products
    put:
      consumes:
      - application/json
      description: Update product details (name, description, price)
      parameters:
      - description: Product ID (UUID)
        in: path
        name: id
        required: true
        type: string
      - description: Updated product details
        in: body
        name: product
        required: true
        schema:
          $ref: '#/definitions/internal_delivery_http_handler.UpdateProductRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Product updated successfully
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Product not found
          schema:
            additionalProperties:
              type: string
            type: object
        "409":
          description: Conflict - product was modified
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Update a product
      tags:
      - Products
  /products/{id}/reviews:
    get:
      consumes:
      - application/json
      description: Get a paginated list of reviews for a specific product. Results
        are cached.
      parameters:
      - description: Product ID (UUID)
        in: path
        name: id
        required: true
        type: string
      - default: 20
        description: Number of items per page (max 100)
        in: query
        name: limit
        type: integer
      - default: 0
        description: Number of items to skip
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Paginated list of reviews
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid product ID
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get reviews for a product
      tags:
      - Reviews
  /reviews:
    post:
      consumes:
      - application/json
      description: Create a new review for a product. Automatically updates product's
        average rating and publishes event.
      parameters:
      - description: Review details
        in: body
        name: review
        required: true
        schema:
          $ref: '#/definitions/internal_delivery_http_handler.CreateReviewRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Review created successfully
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid request body or product not found
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Product not found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Create a new review
      tags:
      - Reviews
  /reviews/{id}:
    delete:
      consumes:
      - application/json
      description: Soft delete a review. Automatically recalculates product's average
        rating and publishes event.
      parameters:
      - description: Review ID (UUID)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: Review deleted successfully
        "400":
          description: Invalid review ID
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Review not found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Delete a review
      tags:
      - Reviews
    put:
      consumes:
      - application/json
      description: Update review details. Automatically recalculates product's average
        rating and publishes event.
      parameters:
      - description: Review ID (UUID)
        in: path
        name: id
        required: true
        type: string
      - description: Updated review details
        in: body
        name: review
        required: true
        schema:
          $ref: '#/definitions/internal_delivery_http_handler.UpdateReviewRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Review updated successfully
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Review not found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Update a review
      tags:
      - Reviews
schemes:
- http
- https
swagger: "2.0"
tags:
- description: Product management endpoints
  name: Products
- description: Review management endpoints
  name: Reviews